Add ContinuationEntry and ContinuationWrapper classes in separate files.
