+++ src/hotspot/cpu/aarch64/continuationHelper_aarch64.inline.hpp
@@ -27,9 +27,25 @@
 
 #include "runtime/continuationHelper.hpp"
 
+#include "runtime/continuationEntry.hpp"
 #include "runtime/frame.inline.hpp"
 #include "runtime/registerMap.hpp"
 #include "utilities/macros.hpp"
+
+inline frame ContinuationEntry::to_frame() const {
+  static CodeBlob* cb = CodeCache::find_blob(entry_pc());
+  return frame(entry_sp(), entry_sp(), entry_fp(), entry_pc(), cb);
+}
+
+inline intptr_t* ContinuationEntry::entry_fp() const {
+  return (intptr_t*)((address)this + size());
+}
+
+inline void ContinuationEntry::update_register_map(RegisterMap* map) const {
+  intptr_t** fp = (intptr_t**)(bottom_sender_sp() - frame::sender_sp_offset);
+  frame::update_map_with_saved_link(map, fp);
+}
+
 
 template<typename FKind>
 static inline intptr_t** link_address(const frame& f) {
@@ -70,15 +86,6 @@ inline void ContinuationHelper::update_register_map(const frame& f, RegisterMap*
   frame::update_map_with_saved_link(map, link_address<FKind>(f));
 }
 
-intptr_t* ContinuationEntry::entry_fp() const {
-  return (intptr_t*)((address)this + size());
-}
-
-void ContinuationEntry::update_register_map(RegisterMap* map) const {
-  intptr_t** fp = (intptr_t**)(bottom_sender_sp() - frame::sender_sp_offset);
-  frame::update_map_with_saved_link(map, fp);
-}
-
 inline void ContinuationHelper::update_register_map_with_callee(const frame& f, RegisterMap* map) {
   frame::update_map_with_saved_link(map, ContinuationHelper::Frame::callee_link_address(f));
 }
@@ -87,18 +94,13 @@ inline void ContinuationHelper::push_pd(const frame& f) {
   *(intptr_t**)(f.sp() - frame::sender_sp_offset) = f.fp();
 }
 
-frame ContinuationEntry::to_frame() const {
-  static CodeBlob* cb = CodeCache::find_blob(entry_pc());
-  return frame(entry_sp(), entry_sp(), entry_fp(), entry_pc(), cb);
-}
 
-
-void ContinuationHelper::set_anchor_to_entry_pd(JavaFrameAnchor* anchor, ContinuationEntry* entry) {
+inline void ContinuationHelper::set_anchor_to_entry_pd(JavaFrameAnchor* anchor, ContinuationEntry* entry) {
   anchor->set_last_Java_fp(entry->entry_fp());
 }
 
 #ifdef ASSERT
-void ContinuationHelper::set_anchor_pd(JavaFrameAnchor* anchor, intptr_t* sp) {
+inline void ContinuationHelper::set_anchor_pd(JavaFrameAnchor* anchor, intptr_t* sp) {
   intptr_t* fp = *(intptr_t**)(sp - frame::sender_sp_offset);
   anchor->set_last_Java_fp(fp);
 }
@@ -117,11 +119,11 @@ inline intptr_t** ContinuationHelper::Frame::callee_link_address(const frame& f)
   return (intptr_t**)(f.sp() - frame::sender_sp_offset);
 }
 
-address* ContinuationHelper::Frame::return_pc_address(const frame& f) {
+inline address* ContinuationHelper::Frame::return_pc_address(const frame& f) {
   return (address*)(f.real_fp() - 1);
 }
 
-address* ContinuationHelper::InterpretedFrame::return_pc_address(const frame& f) {
+inline address* ContinuationHelper::InterpretedFrame::return_pc_address(const frame& f) {
   return (address*)(f.fp() + frame::return_addr_offset);
 }
