+++ src/hotspot/cpu/arm/continuationHelper_arm.inline.hpp
@@ -27,23 +27,24 @@
 
 #include "runtime/continuationHelper.hpp"
 
+#include "runtime/continuationEntry.hpp"
 #include "runtime/frame.inline.hpp"
 #include "runtime/registerMap.hpp"
 #include "utilities/macros.hpp"
 
 // TODO: Implement
 
-frame ContinuationEntry::to_frame() const {
+inline frame ContinuationEntry::to_frame() const {
   Unimplemented();
   return frame();
 }
 
-intptr_t* ContinuationEntry::entry_fp() const {
+inline intptr_t* ContinuationEntry::entry_fp() const {
   Unimplemented();
   return nullptr;
 }
 
-void ContinuationEntry::update_register_map(RegisterMap* map) const {
+inline void ContinuationEntry::update_register_map(RegisterMap* map) const {
   Unimplemented();
 }
 
@@ -76,12 +77,12 @@ inline void ContinuationHelper::push_pd(const frame& f) {
   Unimplemented();
 }
 
-void ContinuationHelper::set_anchor_to_entry_pd(JavaFrameAnchor* anchor, ContinuationEntry* cont) {
+inline void ContinuationHelper::set_anchor_to_entry_pd(JavaFrameAnchor* anchor, ContinuationEntry* cont) {
   Unimplemented();
 }
 
 #ifdef ASSERT
-void ContinuationHelper::set_anchor_pd(JavaFrameAnchor* anchor, intptr_t* sp) {
+inline void ContinuationHelper::set_anchor_pd(JavaFrameAnchor* anchor, intptr_t* sp) {
   Unimplemented();
 }
