+++ src/hotspot/cpu/x86/sharedRuntime_x86_64.cpp
@@ -45,6 +45,7 @@
 #include "oops/compiledICHolder.hpp"
 #include "oops/klass.inline.hpp"
 #include "prims/methodHandles.hpp"
+#include "runtime/continuationEntry.hpp"
 #include "runtime/continuation.hpp"
 #include "runtime/jniHandles.hpp"
 #include "runtime/safepointMechanism.hpp"
